requirements.txt
twilio
openai
flask
python-dotenv